Use static_cast to eliminate MSVC conversion warning

+++ OPHD/UI/Core/RadioButtonGroup.cpp
@@ -19,7 +19,7 @@ RadioButtonGroup::RadioButtonGroup(std::vector<ButtonInfo> buttonInfos)
 	for(auto &buttonInfo : buttonInfos)
 	{
 		NAS2D::Vector<int> offset = {0, 13};
-		offset.y = mRadioButtons.size() * offset.y;
+		offset.y = static_cast<int>(mRadioButtons.size()) * offset.y;
 
 		auto &button = mRadioButtons.emplace_back(this, buttonInfo.name, buttonInfo.delegate);
 		button.visible(visible());
